fix: Initialized user in DeckPicker
The user was initialized in Game, but since the landing activity is DeckPicker,
that caused a problem when checking the leaderboard from the DeckPicker. Since
at that point the user has not been initialized, the connection to the database
was faulty. Before checking the leaderboard, the app updates the number of points,
this is done by getting the node corresponding to the user, but since it has not
been created, the app puts the value of points below the node users, that way
when retrieving the children of that node, some of them won't be properly formed.